Feature/snapshot (#17)
* Bump to minimum iOS 12.4.

Bump minimum deployment target from iOS 12.2 to 12.4. All our old devices at Appamajigger are at 12.4.6.

* Plumb button for saving photo of board.

* Apple's SVG for camera icon

Exported from Apple's SF Symbols app.

* camera button icon

For iOS 13 or later, can use built-in iOS "SF Images", but for iOS 12 or earlier you have to supply your own icons. These are created from artwork supplied by Apple -- see Apple-camera.svg in this same directory.

* Yellow camera icons

Made versions of the camera icons that are the same yellow as the text used in the Go Stone Art app.

* Work in progress on snapshot button

* Camera button for compact width

The layout works for iPhones and other compact width size class devices, but the camera icon is too small for iPads.

* 48x48 version of camera icon

Need larger version of icon for iPads.

* Two size classes of snapshot button

Depending on size class, one of two different camera buttons will appear. A 24x24 icon for compact devices or a 48x48 icon for regular devices.

* iPad needs more spacing between buttons.

* First shot at BoardView.savePhoto()

* Clean up name of small camera icon.

It was bothering me that the large version of the icon was called camera48, but the small version was not called camera24.

* Completion handler for UIImageWriteToSavedPhotosAlbum()

* More meaningful name for completion selector

* Make completion selector a private function.

* Completion selector parameter name cleanup

* Clean up UIImageWriteToSavedPhotosAlbum() call.

* Extension for alerts with OK button

* Dismiss alert when app resigns active.

* Alert now dismisses itself when app goes inactive.

* When successfully save photo limit time alert is shown.

* Offer to take user to Settings.

* Clean up BoardView code for saving photos.